requirements.txt
fastapi
uvicorn
pyjwt
passlib
cryptography
python-dotenv
python-multipart
sqlalchemy
slowapi